+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Fixed a bug, due to which the installed engine block would slowly rotate, when car engine block is repeatedly enabled and disabled
 - Fixed MOP not starting, if the "envelope(xxxxx)Paid" was set to true manually with MSCEditor
+- Fixed the starting beer case sometimes not being hooked with ItemBehaviour class
 
 ## 3.8.6 (08.09.2022)